+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 - DatabaseConnection.rollbackTransaction() and commitTransaction() no longer ignore exceptions, now throw SQLException.
 ### is.codion.swing.common.ui
 - FilterColumnConditionPanel.Builder.operatorCaptions() added.
+- ColumnConditionPanel.identifier() removed along with type parameter.
 ### is.codion.framework.db
 - EntityConnection.rollbackTransaction() and commitTransaction() no longer ignore exceptions, now throw DatabaseException.
 - EntityConnection.transaction() added along with Transactional and TransactionalResult.